+++ commerce/payment/src/main/java/ru/yandex/practicum/model/Payment.java
@@ -3,6 +3,7 @@ package ru.yandex.practicum.model;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Entity
@@ -18,11 +19,11 @@ public class Payment {
     @Column(name = "payment_id", nullable = false)
     private UUID paymentId;
     @Column(name = "total_payment")
-    private double totalPayment;
+    private BigDecimal totalPayment;
     @Column(name = "total_delivery")
-    private double deliveryTotal;
+    private BigDecimal deliveryTotal;
     @Column(name = "total_fee")
-    private double feeTotal;
+    private BigDecimal feeTotal;
     @Column(name = "payment_status")
     @Enumerated(EnumType.STRING)
     private PaymentStatus paymentStatus;
