+++ commerce/payment/src/main/java/ru/yandex/practicum/service/PaymentService.java
@@ -3,6 +3,7 @@ package ru.yandex.practicum.service;
 import ru.yandex.practicum.dto.ProductDto;
 import ru.yandex.practicum.model.Payment;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
@@ -12,7 +13,7 @@ public interface PaymentService {
 
     Payment createPayment(Payment payment);
 
-    double calculateTotalCost(double deliveryPrice, double productsPrice);
+    BigDecimal calculateTotalCost(BigDecimal deliveryPrice, BigDecimal productsPrice);
 
     void successPayment(UUID orderId);
